% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acp_sernanp}
\alias{acp_sernanp}
\title{Areas de Conservacion Privada (ACP) reconocidas por SERNANP}
\format{
Un data frame (\code{tbl_df}) con informacion de:
\describe{
\item{departamento}{Nombre del departamento donde se ubica el area.}
\item{area_de_conservacion}{Nombre oficial del Area de Conservacion Privada.}
\item{base_legal}{Norma legal de reconocimiento o establecimiento.}
\item{sup_ha}{Superficie total del area en hectareas (ha).}
\item{fecha_promulgacion}{Fecha oficial de promulgacion o reconocimiento.}
}
}
\source{
Servicio Nacional de Areas Naturales Protegidas por el Estado (SERNANP) / INEI.
}
\usage{
acp_sernanp
}
\description{
Conjunto de datos con el registro de las Areas de Conservacion Privada (ACP)
en el territorio peruano, reconocidas mediante resolucion ministerial.
}
\keyword{datasets}
