% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acr_sernanp}
\alias{acr_sernanp}
\title{Areas de Conservacion Regional (ACR) reconocidas por SERNANP}
\format{
Un data frame con variables de ubicacion, nombre del ACR, base legal y superficie en hectareas.
}
\source{
Servicio Nacional de Areas Naturales Protegidas por el Estado (SERNANP) / INEI.
}
\usage{
acr_sernanp
}
\description{
Informacion sobre las Areas de Conservacion Regional (ACR) establecidas en el Peru
para la proteccion de ecosistemas y biodiversidad a nivel subnacional.
}
\keyword{datasets}
