% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cap_pot_bosq_amaz}
\alias{cap_pot_bosq_amaz}
\title{Capacidad de uso mayor y potencial de los bosques humedos amazonicos}
\format{
Un data frame con clasificacion de suelos y capacidad forestal/de proteccion en hectareas.
}
\source{
Instituto Nacional de Estadistica e Informatica (INEI) / MINAM.
}
\usage{
cap_pot_bosq_amaz
}
\description{
Informacion sobre la capacidad potencial del suelo y cobertura de bosques humedos en la Amazonia peruana.
}
\keyword{datasets}
