% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flora_fauna_2013_2020}
\alias{flora_fauna_2013_2020}
\title{Especies de Flora y Fauna Amenazadas (2013-2020)}
\format{
Un data frame con registros anuales por grupo taxonomico y categoria de amenaza.
}
\source{
SERFOR / Ministerio del Ambiente (MINAM) / INEI.
}
\usage{
flora_fauna_2013_2020
}
\description{
Registro historico del numero de especies de flora y fauna silvestre amenazadas
segun categorias de conservacion (Vulnerable, En Peligro, En Peligro Critico).
}
\keyword{datasets}
