% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flora_fauna_ende}
\alias{flora_fauna_ende}
\title{Especies de Flora y Fauna Endemicas}
\format{
Un data frame con grupos taxonomicos (aves, mamiferos, anfibios, reptiles, plantas) y conteo de especies endemicas.
}
\source{
Ministerio del Ambiente (MINAM) / INEI.
}
\usage{
flora_fauna_ende
}
\description{
Numero de especies endemicas registradas en el Peru por grupo taxonomico.
}
\keyword{datasets}
