% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humedad_rel}
\alias{humedad_rel}
\title{Humedad Relativa Promedio Anual segun Departamento}
\format{
Un data frame con departamentos y valores anuales de humedad relativa (\%).
}
\source{
Servicio Nacional de Meteorologia e Hidrologia del Peru (SENAMHI) / INEI.
}
\usage{
humedad_rel
}
\description{
Serie historica de humedad relativa promedio anual (\%) registrada en estaciones
meteorologicas representativas de cada departamento del Peru.
}
\keyword{datasets}
