% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{islas_islotes}
\alias{islas_islotes}
\title{Islas e Islotes del litoral peruano}
\format{
Un data frame con nombre, ubicacion departamental y superficie estimada.
}
\source{
Direccion de Hidrografia y Navegacion (DHN) / INEI.
}
\usage{
islas_islotes
}
\description{
Inventario de las principales islas, islotes y puntas guaneras a lo largo de la costa peruana.
}
\keyword{datasets}
