% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precipitacion}
\alias{precipitacion}
\title{Precipitacion Total Anual segun Departamento}
\format{
Un data frame con departamentos y valores anuales de precipitacion en mm.
}
\source{
Servicio Nacional de Meteorologia e Hidrologia del Peru (SENAMHI) / INEI.
}
\usage{
precipitacion
}
\description{
Serie historica de precipitacion total anual acumulada (milimetros, mm) registrada
en las principales estaciones meteorologicas por departamento.
}
\keyword{datasets}
