% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinia_api.R
\name{print.sinia_estadistica}
\alias{print.sinia_estadistica}
\title{Imprimir objeto de estadística SINIA}
\usage{
\method{print}{sinia_estadistica}(x, ...)
}
\arguments{
\item{x}{Objeto de clase \code{sinia_estadistica}.}

\item{...}{Argumentos adicionales pasados a otros métodos.}
}
\value{
Retorna el objeto \code{x} de forma invisible.
}
\description{
Imprimir objeto de estadística SINIA
}
