% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prom_mensual_pm_menor_10mcr}
\alias{prom_mensual_pm_menor_10mcr}
\title{Promedio Mensual de Material Particulado PM10 en Lima Metropolitana}
\format{
Un data frame con registros mensuales, estacion de monitoreo y concentracion en ug/m3.
}
\source{
DIGESA / SENAMHI / OEFA / INEI.
}
\usage{
prom_mensual_pm_menor_10mcr
}
\description{
Concentraciones promedio mensuales de material particulado con diametro inferior
a 10 micras (PM10 en ug/m3) monitoreadas en diferentes estaciones de Lima Metropolitana.
}
\keyword{datasets}
