% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rios_frontera}
\alias{rios_frontera}
\title{Rios Principales y de Cuencas Transfronterizas del Peru}
\format{
Un data frame con nombre del rio, vertiente hidrografica, longitud y cuenca.
}
\source{
Autoridad Nacional del Agua (ANA) / INEI.
}
\usage{
rios_frontera
}
\description{
Caracteristicas hidrogaficas de los principales rios y sistemas hidricos compartidos en fronteras.
}
\keyword{datasets}
