% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinia_api.R
\name{sinia_api}
\alias{sinia_api}
\title{Interfaz para Estadísticas Ambientales del SINIA / MINAM}
\description{
Funciones para listar indicadores, buscar, consultar fichas técnicas
y descargar conjuntos de datos del Sistema Nacional de Información Ambiental (SINIA).

El Sistema Nacional de Información Ambiental (SINIA) es un sistema funcional
del Sistema Nacional de Gestión Ambiental (SNGA), que comprende un conjunto de
principios, normas, procedimientos, técnicas e instrumentos a fin de facilitar
el acceso y uso de la información ambiental que las entidades que lo conforman
generan o poseen, en el ámbito de sus respectivas competencias.
}
\references{
Ministerio del Ambiente del Perú (MINAM). Sistema Nacional de Información Ambiental (SINIA).
\url{https://sinia.minam.gob.pe/}
}
\keyword{internal}
