% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{siniaR-datasets}
\alias{siniaR-datasets}
\title{Datos Ambientales del Peru (INEI / SERNANP / SENAMHI)}
\description{
Conjuntos de datos ambientales oficiales compilados a partir de los anuarios
estadisticos del Instituto Nacional de Estadistica e Informatica (INEI) y organismos
adscritos al Ministerio del Ambiente (SERNANP, SENAMHI, OEFA).
}
