% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sup_dep_sup_bha}
\alias{sup_dep_sup_bha}
\title{Superficie Departamental y Superficie con Bosque Humedo Amazonico}
\format{
Un data frame con departamento, superficie departamental (ha) y superficie de bosque (ha y \%).
}
\source{
Programa Nacional de Conservacion de Bosques (PNCBMCC) / INEI.
}
\usage{
sup_dep_sup_bha
}
\description{
Distribucion de la superficie de bosque humedo amazonico (BHA) y su porcentaje
respecto a la superficie territorial total de cada departamento.
}
\keyword{datasets}
