% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{superficie_bha}
\alias{superficie_bha}
\title{Superficie de Bosque Humedo Amazonico}
\format{
Un data frame con registros anuales de superficie en hectareas.
}
\source{
MINAM / PNCBMCC / INEI.
}
\usage{
superficie_bha
}
\description{
Serie historica de la cobertura de bosque humedo amazonico en el territorio nacional.
}
\keyword{datasets}
