% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperatura}
\alias{temperatura}
\title{Temperatura del Aire Promedio Anual segun Departamento}
\format{
Un data frame con departamento y serie de temperaturas promedio anuales (°C).
}
\source{
Servicio Nacional de Meteorologia e Hidrologia del Peru (SENAMHI) / INEI.
}
\usage{
temperatura
}
\description{
Serie historica de temperatura media anual del aire (grados Celsius, °C) registrada
en estaciones meteorologicas de capitales departamentales.
}
\keyword{datasets}
