% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperatura_maxima}
\alias{temperatura_maxima}
\title{Temperatura Maxima Promedio Anual segun Departamento}
\format{
Un data frame con departamento y serie de temperaturas maximas anuales (°C).
}
\source{
Servicio Nacional de Meteorologia e Hidrologia del Peru (SENAMHI) / INEI.
}
\usage{
temperatura_maxima
}
\description{
Serie historica de temperatura maxima promedio anual (grados Celsius, °C) registrada
por departamento.
}
\keyword{datasets}
