% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xlsx_links}
\alias{xlsx_links}
\title{Catalogo de Enlaces a Archivos Excel Tematicos del INEI}
\format{
Un data frame con titulo de la tabla y URL de descarga oficial en formato \code{.xlsx}.
}
\source{
Instituto Nacional de Estadistica e Informatica (INEI).
}
\usage{
xlsx_links
}
\description{
Tabla de referencia con los titulos tematicos y enlaces directos de descarga
a las tablas estadisticas del anuario ambiental del INEI.
}
\keyword{datasets}
