% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinia_api.R
\name{print.sinia_ficha}
\alias{print.sinia_ficha}
\title{Imprimir ficha técnica de SINIA}
\usage{
\method{print}{sinia_ficha}(x, ...)
}
\arguments{
\item{x}{Objeto de clase \code{sinia_ficha}.}

\item{...}{Argumentos adicionales pasados a otros métodos.}
}
\value{
Retorna el objeto \code{x} de forma invisible.
}
\description{
Imprimir ficha técnica de SINIA
}
\keyword{internal}
