% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinia_api.R
\name{sinia_api}
\alias{sinia_api}
\title{Interfaz para Estadísticas Ambientales del SINIA / MINAM}
\description{
Funciones para listar indicadores, buscar, consultar fichas técnicas
y descargar conjuntos de datos del Sistema Nacional de Información Ambiental (SINIA).
}
\keyword{internal}
