% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinia_api.R
\name{sinia_estadistica}
\alias{sinia_estadistica}
\title{Obtener datos y ficha tecnica completa de una estadistica del SINIA}
\usage{
sinia_estadistica(id, pivot = c("wide", "long"))
}
\arguments{
\item{id}{Identificador numerico de la estadistica (ej. \code{1} para temperatura).}

\item{pivot}{Formato de salida de los datos (\code{"wide"} o \code{"long"}).}
}
\value{
Un objeto de clase \code{sinia_estadistica} que contiene:
\itemize{
\item \code{ficha}: Objeto \code{sinia_ficha} con los metadatos.
\item \code{datos}: \link[tibble:tibble]{tibble::tibble} con los datos tabulares.
}
}
\description{
Descarga tanto los metadatos de la ficha tecnica como el conjunto de datos
estructurado en un unico objeto \code{sinia_estadistica}.
}
\examples{
\dontrun{
est <- sinia_estadistica(1)
est$ficha
head(est$datos)
}
}
